Since Files in ReplayFileAppender can be null (if their name entry should be removed from the ZIP), a null check was added when calling delete()

+++ src/main/java/eu/crushedpixel/replaymod/registry/ReplayFileAppender.java
@@ -95,8 +95,11 @@ public class ReplayFileAppender extends Thread {
 
                         //delete all written files
                         for(Pair<File, String> p : filesToMove.get(replayFile)) {
-                            p.getLeft().delete();
+                            if(p.getLeft() != null) {
+                                p.getLeft().delete();
+                            }
                         }
+
                         filesToMove.removeAll(replayFile);
                     } catch(Exception e) {
                         e.printStackTrace();
